target a specific GATE release
unfortunately `project.version` changes down the inheritance hierarchy so using that to reference a GATE version breaks when the plugin isn't one of the ones that used to be in core. We now reference a specific GATE version but allow this to be easily overridden in plugins that inherit from this pom.